+++ Modules/Blog/app/Providers/BlogServiceProvider.php
@@ -12,6 +12,7 @@ namespace Modules\Blog\Providers;
 
 use Illuminate\Support\Facades\Gate;
 use Livewire\Livewire;
+use Modules\Blog\Console\ArticleCategoryCommand;
 use Modules\Blog\Console\ArticleVerifyCommand;
 use Modules\Blog\Console\DownloadWpImagesCommand;
 use Modules\Blog\Console\ImportWordPressCommand;
@@ -33,7 +34,7 @@ class BlogServiceProvider extends BaseModuleServiceProvider
     public function boot(): void
     {
         $this->bootModule();
-        $this->commands([ImportWordPressCommand::class, DownloadWpImagesCommand::class, MigrateContentImagesCommand::class, YouTubeSummarizeCommand::class, ArticleVerifyCommand::class]);
+        $this->commands([ImportWordPressCommand::class, DownloadWpImagesCommand::class, MigrateContentImagesCommand::class, YouTubeSummarizeCommand::class, ArticleVerifyCommand::class, ArticleCategoryCommand::class]);
 
         Gate::policy(Article::class, ArticlePolicy::class);
         Gate::policy(Comment::class, CommentPolicy::class);